[Finder] Moved Shell & Command classes to Shell namespace.
